feat(summary): extract leaked data volume
