+++ internal/summary/client_test.go
@@ -3,11 +3,9 @@ package summary
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
-	"strings"
 	"testing"
 	"time"
 )
@@ -72,253 +70,6 @@ func TestClientGenerate_returnsSummary_whenEndpointSupportsMinimalChatCompletion
 	}
 	if _, exists := body["max_completion_tokens"]; exists {
 		t.Fatalf("max_completion_tokens must be omitted: %#v", body)
-	}
-}
-
-func TestClientAnalyzeArticleSendsFullBody_andReturnsImpactSignals(t *testing.T) {
-	// Given an OpenAI-compatible endpoint and a body marker that appears only at the article end.
-	requests := make(chan map[string]any, 1)
-	upstream := httptest.NewServer(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		var body map[string]any
-		if err := json.NewDecoder(request.Body).Decode(&body); err != nil {
-			t.Errorf("decode request: %v", err)
-		}
-		requests <- body
-		writer.Header().Set("Content-Type", "application/json")
-		_, _ = writer.Write([]byte(`{"id":"chatcmpl-test","object":"chat.completion","created":1,"model":"test-model","choices":[{"index":0,"message":{"role":"assistant","content":"{\"summary\":\"Incident summary\",\"attack_method\":\"Supply chain\",\"threat_actor\":\"Unknown\",\"actor_country\":\"\",\"target_sector\":\"IT\",\"victim_count\":15000,\"zero_day\":true}"},"finish_reason":"stop"}]}`))
-	}))
-	defer upstream.Close()
-	client, err := NewClient(Config{BaseURL: upstream.URL + "/v1", Model: "test-model", APIKey: "key", Timeout: 2 * time.Second})
-	if err != nil {
-		t.Fatalf("new client: %v", err)
-	}
-	fullBody := "Opening paragraph.\n\nTechnical details.\n\nFULL_BODY_END_MARKER"
-
-	// When the article is analyzed.
-	analysis, err := client.AnalyzeArticle(context.Background(), ArticleRequest{
-		Language: "en", Title: "Supply-chain incident", URL: "https://example.com/story", Body: fullBody,
-	})
-
-	// Then the complete body reaches the endpoint and structured impact signals are parsed.
-	requestBody := <-requests
-	if err != nil {
-		t.Fatalf("analyze article: %v", err)
-	}
-	messages, ok := requestBody["messages"].([]any)
-	if !ok || len(messages) < 2 {
-		t.Fatalf("messages = %#v", requestBody["messages"])
-	}
-	userMessage, ok := messages[1].(map[string]any)
-	content, _ := userMessage["content"].(string)
-	if !ok || !strings.Contains(content, "FULL_BODY_END_MARKER") {
-		t.Fatalf("user message does not contain the complete body: %#v", userMessage)
-	}
-	if analysis.VictimCount != 15000 || !analysis.ZeroDay || analysis.Summary != "Incident summary" {
-		t.Fatalf("analysis = %+v", analysis)
-	}
-}
-
-func TestClientAnalyzeArticleAcceptsJSONCodeFence(t *testing.T) {
-	upstream := httptest.NewServer(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		writer.Header().Set("Content-Type", "application/json")
-		content := "```json\n{\"summary\":\"Incident summary\",\"attack_method\":\"Malware\",\"threat_actor\":\"Unknown\",\"actor_country\":\"\",\"target_sector\":\"Finance\",\"victim_count\":0,\"zero_day\":false}\n```"
-		encodedContent, _ := json.Marshal(content)
-		_, _ = fmt.Fprintf(writer, `{"id":"chatcmpl-test","object":"chat.completion","created":1,"model":"test-model","choices":[{"index":0,"message":{"role":"assistant","content":%s},"finish_reason":"stop"}]}`, encodedContent)
-	}))
-	defer upstream.Close()
-	client, err := NewClient(Config{BaseURL: upstream.URL + "/v1", Model: "test-model", APIKey: "key", Timeout: 2 * time.Second})
-	if err != nil {
-		t.Fatalf("new client: %v", err)
-	}
-
-	analysis, err := client.AnalyzeArticle(context.Background(), ArticleRequest{Language: "en", Title: "Incident", Body: "Body"})
-
-	if err != nil {
-		t.Fatalf("analyze article: %v", err)
-	}
-	if analysis.Summary != "Incident summary" || analysis.TargetSector != "Finance" {
-		t.Fatalf("analysis = %+v", analysis)
-	}
-}
-
-func TestClientAnalyzeArticleTreatsNumericZeroCountryAsUnknown(t *testing.T) {
-	// Given a compatible endpoint that uses numeric zero for an unknown actor country.
-	upstream := httptest.NewServer(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		writer.Header().Set("Content-Type", "application/json")
-		_, _ = writer.Write([]byte(`{"id":"chatcmpl-test","object":"chat.completion","created":1,"model":"test-model","choices":[{"index":0,"message":{"role":"assistant","content":"{\"summary\":\"Security strategy\",\"attack_method\":\"AI-driven attacks\",\"threat_actor\":\"attackers\",\"actor_country\":0,\"target_sector\":\"enterprise security\",\"victim_count\":500,\"zero_day\":false}"},"finish_reason":"stop"}]}`))
-	}))
-	defer upstream.Close()
-	client, err := NewClient(Config{BaseURL: upstream.URL + "/v1", Model: "test-model", APIKey: "key", Timeout: 2 * time.Second})
-	if err != nil {
-		t.Fatalf("new client: %v", err)
-	}
-
-	// When the article response is analyzed.
-	analysis, err := client.AnalyzeArticle(context.Background(), ArticleRequest{Language: "en", Title: "AI security", Body: "Body"})
-
-	// Then the usable analysis is retained and zero becomes an empty optional country.
-	if err != nil {
-		t.Fatalf("analyze article: %v", err)
-	}
-	if analysis.ActorCountry != "" || analysis.Summary != "Security strategy" {
-		t.Fatalf("analysis = %+v", analysis)
-	}
-}
-
-func TestClientAnalyzeArticleUsesUnknownActor_whenResponseLeavesThreatActorEmpty(t *testing.T) {
-	// Given an otherwise complete analysis whose article does not name a threat actor.
-	client := newArticleAnalysisClient(t, `{
-		"summary":"Adobe fixed critical vulnerabilities.",
-		"attack_method":"SQL injection",
-		"threat_actor":"",
-		"actor_country":"",
-		"target_sector":"Enterprise software",
-		"victim_count":0,
-		"zero_day":false
-	}`)
-
-	// When the compatible endpoint response is parsed.
-	analysis, err := client.AnalyzeArticle(context.Background(), ArticleRequest{Language: "ko", Title: "Adobe update", Body: "Body"})
-
-	// Then the unnamed actor is represented as unknown instead of rejecting the whole analysis.
-	if err != nil {
-		t.Fatalf("analyze article: %v", err)
-	}
-	if analysis.ThreatActor != "Unknown" {
-		t.Fatalf("threat actor = %q, want Unknown", analysis.ThreatActor)
-	}
-}
-
-func TestClientAnalyzeArticleJoinsAttackMethods_whenResponseUsesArray(t *testing.T) {
-	// Given a compatible endpoint that returns multiple attack methods as a JSON array.
-	client := newArticleAnalysisClient(t, `{
-		"summary":"DNS security products address several attack techniques.",
-		"attack_method":["피싱","악성코드 콜백","DNS 터널링"],
-		"threat_actor":"",
-		"actor_country":"",
-		"target_sector":"기업",
-		"victim_count":0,
-		"zero_day":false
-	}`)
-
-	// When the compatible endpoint response is parsed.
-	analysis, err := client.AnalyzeArticle(context.Background(), ArticleRequest{Language: "ko", Title: "DNS security", Body: "Body"})
-
-	// Then the list is retained as one display-safe category value.
-	if err != nil {
-		t.Fatalf("analyze article: %v", err)
-	}
-	if analysis.AttackMethod != "피싱, 악성코드 콜백, DNS 터널링" {
-		t.Fatalf("attack method = %q", analysis.AttackMethod)
-	}
-}
-
-func TestClientAnalyzeArticleReadsDamageAmount_whateverShapeTheModelWritesItIn(t *testing.T) {
-	tests := []struct {
-		name  string
-		value string
-		want  int64
-	}{
-		{name: "whole number", value: `1500000`, want: 1_500_000},
-		{name: "float", value: `1.5e6`, want: 1_500_000},
-		{name: "prose amount", value: `"$1.5 million"`, want: 1_500_000},
-		{name: "separated digits", value: `"USD 190,000,000"`, want: 190_000_000},
-		{name: "no figure", value: `"unknown"`, want: 0},
-		{name: "absent", value: `null`, want: 0},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			// Given an analysis whose damage figure arrives in one of the shapes models use.
-			client := newArticleAnalysisClient(t, `{
-				"summary":"Attackers drained a bridge.",
-				"attack_method":"Financial / Crypto",
-				"threat_actor":"Unknown",
-				"actor_country":"",
-				"target_sector":"Finance",
-				"victim_count":0,
-				"damage_usd":`+test.value+`,
-				"zero_day":false
-			}`)
-
-			// When the response is parsed.
-			analysis, err := client.AnalyzeArticle(context.Background(), ArticleRequest{Language: "en", Title: "Bridge drained", Body: "Body"})
-
-			// Then the amount reaches severity as whole US dollars, and an unusable figure
-			// reads as no stated damage instead of discarding the whole analysis.
-			if err != nil {
-				t.Fatalf("analyze article: %v", err)
-			}
-			if analysis.DamageUSD != test.want {
-				t.Fatalf("damage = %d, want %d", analysis.DamageUSD, test.want)
-			}
-		})
-	}
-}
-
-func TestClientAnalyzeArticleKeepsPatchStateToThreeValues(t *testing.T) {
-	tests := []struct {
-		name  string
-		value string
-		want  string
-	}{
-		{name: "fix released", value: `"yes"`, want: "yes"},
-		{name: "no fix yet", value: `"no"`, want: "no"},
-		{name: "not stated", value: `"unknown"`, want: ""},
-		{name: "wording instead of the enum", value: `"Patched"`, want: "yes"},
-		{name: "boolean true", value: `true`, want: "yes"},
-		// A model answering false about an article that never mentioned a fix would
-		// otherwise raise the article's severity for a claim it never made.
-		{name: "boolean false", value: `false`, want: ""},
-		{name: "prose", value: `"a fix is expected next week"`, want: ""},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			// Given an analysis carrying one of the shapes models write for patch state.
-			client := newArticleAnalysisClient(t, `{
-				"summary":"A critical flaw was disclosed.",
-				"attack_method":"Vulnerability Disclosure",
-				"threat_actor":"None",
-				"actor_country":"",
-				"target_sector":"Technology",
-				"victim_count":0,
-				"damage_usd":0,
-				"patch_available":`+test.value+`,
-				"zero_day":false
-			}`)
-
-			// When the response is parsed.
-			analysis, err := client.AnalyzeArticle(context.Background(), ArticleRequest{Language: "en", Title: "Advisory", Body: "Body"})
-
-			// Then only a stated fix state survives; everything else means nothing stated.
-			if err != nil {
-				t.Fatalf("analyze article: %v", err)
-			}
-			if analysis.PatchAvailable != test.want {
-				t.Fatalf("patch available = %q, want %q", analysis.PatchAvailable, test.want)
-			}
-		})
-	}
-}
-
-func TestClientAnalyzeArticleRejectsNegativeDamage(t *testing.T) {
-	// Given an otherwise complete analysis reporting a negative loss.
-	client := newArticleAnalysisClient(t, `{
-		"summary":"Incident summary",
-		"attack_method":"Ransomware",
-		"threat_actor":"Unknown",
-		"actor_country":"",
-		"target_sector":"Healthcare",
-		"victim_count":0,
-		"damage_usd":-5000,
-		"zero_day":false
-	}`)
-
-	// When the response is parsed.
-	_, err := client.AnalyzeArticle(context.Background(), ArticleRequest{Language: "en", Title: "Incident", Body: "Body"})
-
-	// Then it is rejected rather than lowering the article's severity with nonsense.
-	if !errors.Is(err, ErrInvalidResponse) {
-		t.Fatalf("error = %v, want %v", err, ErrInvalidResponse)
 	}
 }
